Backend/requirements.txt
Django==5.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1  
django-environ==0.11.2 

Django>=4.0
reportlab
